+++ examples/fsi/falling_rotating_rigid_squares_2d.jl
@@ -115,11 +115,11 @@ boundary_model_structure_2 = BoundaryModelDummyParticles(hydrodynamic_densities_
                                                          fluid_smoothing_kernel,
                                                          fluid_smoothing_length)
 
-structure_system_1 = RigidSPHSystem(square1;
+structure_system_1 = RigidBodySystem(square1;
                                     boundary_model=boundary_model_structure_1,
                                     acceleration=(0.0, -gravity),
                                     particle_spacing=structure_particle_spacing)
-structure_system_2 = RigidSPHSystem(square2;
+structure_system_2 = RigidBodySystem(square2;
                                     boundary_model=boundary_model_structure_2,
                                     acceleration=(0.0, -gravity),
                                     particle_spacing=structure_particle_spacing)