+++ src/general/neighborhood_search.jl
@@ -52,7 +52,7 @@ end
 
 # -- Boundary models
 @inline function compact_support(system::Union{TotalLagrangianSPHSystem,
-                                               RigidSPHSystem,
+                                               RigidBodySystem,
                                                WallBoundarySystem},
                                  neighbor)
     return compact_support(system, system.boundary_model, neighbor)
@@ -77,7 +77,7 @@ end
     return compact_support(smoothing_kernel, smoothing_length)
 end
 
-@inline function compact_support(system::RigidSPHSystem, model::Nothing, neighbor)
+@inline function compact_support(system::RigidBodySystem, model::Nothing, neighbor)
     # Rigid interactions without a boundary model are currently not modeled.
     # This neighborhood search exists only to keep the semidiscretization structure uniform.
     return zero(eltype(system))
@@ -207,7 +207,7 @@ end
 function update_nhs!(neighborhood_search,
                      system::AbstractFluidSystem,
                      neighbor::Union{AbstractFluidSystem, TotalLagrangianSPHSystem,
-                                     RigidSPHSystem},
+                                     RigidBodySystem},
                      u_system, u_neighbor, semi)
     # The current coordinates of fluids and structures change over time
     update!(neighborhood_search,
@@ -278,7 +278,7 @@ end
 
 function update_nhs!(neighborhood_search,
                      system::OpenBoundarySystem,
-                     neighbor::RigidSPHSystem,
+                     neighbor::RigidBodySystem,
                      u_system, u_neighbor, semi)
     update!(neighborhood_search,
             current_coordinates(u_system, system),
@@ -297,7 +297,7 @@ function update_nhs!(neighborhood_search,
 end
 
 function update_nhs!(neighborhood_search,
-                     system::RigidSPHSystem,
+                     system::RigidBodySystem,
                      neighbor::OpenBoundarySystem,
                      u_system, u_neighbor, semi)
     update!(neighborhood_search,
@@ -335,7 +335,7 @@ function update_nhs!(neighborhood_search,
 end
 
 function update_nhs!(neighborhood_search,
-                     system::TotalLagrangianSPHSystem, neighbor::RigidSPHSystem,
+                     system::TotalLagrangianSPHSystem, neighbor::RigidBodySystem,
                      u_system, u_neighbor, semi)
     # The current coordinates of both systems change over time.
     update!(neighborhood_search,
@@ -364,9 +364,9 @@ function update_nhs!(neighborhood_search,
 end
 
 function update_nhs!(neighborhood_search,
-                     system::RigidSPHSystem,
+                     system::RigidBodySystem,
                      neighbor::Union{AbstractFluidSystem, TotalLagrangianSPHSystem,
-                                     RigidSPHSystem},
+                                     RigidBodySystem},
                      u_system, u_neighbor, semi)
     # The current coordinates of fluids and structures change over time.
     update!(neighborhood_search,
@@ -376,7 +376,7 @@ function update_nhs!(neighborhood_search,
 end
 
 function update_nhs!(neighborhood_search,
-                     system::RigidSPHSystem, neighbor::WallBoundarySystem,
+                     system::RigidBodySystem, neighbor::WallBoundarySystem,
                      u_system, u_neighbor, semi)
     # The current coordinates of structures change over time.
     # Boundary coordinates only change over time when `neighbor.ismoving[]`.
@@ -446,7 +446,7 @@ end
 # Explicitly define this method to avoid ambiguity with the generic no-op method below.
 function update_nhs!(neighborhood_search,
                      system::WallBoundarySystem{<:BoundaryModelDummyParticles},
-                     neighbor::RigidSPHSystem,
+                     neighbor::RigidBodySystem,
                      u_system, u_neighbor, semi)
     # Don't update. This NHS is never used.
     return neighborhood_search
@@ -489,7 +489,7 @@ end
 # -- Combinations that are never used
 function update_nhs!(neighborhood_search,
                      system::WallBoundarySystem,
-                     neighbor::Union{AbstractFluidSystem, RigidSPHSystem},
+                     neighbor::Union{AbstractFluidSystem, RigidBodySystem},
                      u_system, u_neighbor, semi)
     # Don't update. This NHS is never used.
     return neighborhood_search