+++ test/schemes/fluid/surface_normal_sph.jl
@@ -56,7 +56,7 @@ function create_rigid_boundary_system(coordinates, particle_spacing, state_equat
                                               correction=nothing,
                                               reference_particle_spacing=particle_spacing)
 
-    return RigidSPHSystem(wall_system.initial_condition;
+    return RigidBodySystem(wall_system.initial_condition;
                           boundary_model=rigid_model,
                           acceleration=ntuple(_ -> 0.0, NDIMS))
 end